+++ cocos2d/CCSpriteBatchNode.h
@@ -144,7 +144,7 @@
 -(void) removeSpriteFromAtlas:(CCSprite*)sprite;
 
 -(NSUInteger) rebuildIndexInOrder:(CCSprite*)parent atlasIndex:(NSUInteger)index;
--(NSUInteger) atlasIndexForChild:(CCSprite*)sprite atZ:(int)z;
+-(NSUInteger) atlasIndexForChild:(CCSprite*)sprite atZ:(NSInteger)z;
 /* Sprites use this to start sortChildren, don't call this manually */
 - (void) reorderBatch;
