+++ cocos2d/Support/CCFileUtils.h
@@ -49,7 +49,7 @@
  @returns the size of the allocated buffer
  @since v0.99.5
  */
-int ccLoadFileIntoMemory(const char *filename, unsigned char **out);
+NSInteger ccLoadFileIntoMemory(const char *filename, unsigned char **out);
 
 
 /** removes the HD suffix from a path
